requirements.txt
accelerate
bitsandbytes
datasets
joblib
matplotlib
mediapipe
opencv-contrib-python-headless
opencv-python
numpy<2.0.0
pillow
PyYAML
seaborn
scikit-learn
setuptools
timm
torch
torchvision
tqdm
transformers
ultralytics